Made Objects exist
Added the objects class, with an x, y, and a sprite id. You can make instances of objects with those variables, there is an std::vector to hold the objects (which will use pointers later on), and the program can display all the objects in that vector onto the screen.